requirements-workshop.txt
# Workshop dependencies (install inside a virtual environment).
# Usage: pip install -r requirements-workshop.txt

google-adk>=1.0.0
# Optional — for `adk eval` (heavier deps): pip install "google-adk[eval]"
jupyter>=1.1.0
ipykernel>=6.29.0
pytest>=8.0.0
requests>=2.28.0
